+++ helpers/element_creaters/createWinnerTitle.js
@@ -1,12 +1,12 @@
-import {createNewElement} from "./createNewElement.js";
-
-
-export function winner(name) {
-    const $winnerTitle = createNewElement('div', 'winnerTitle')
-    if (name) {
-        $winnerTitle.innerText = `${name} wins!`
-        return $winnerTitle
-    }
-    $winnerTitle.innerText = 'draw'
-    return $winnerTitle
-}
+// import {createNewElement} from "./createNewElement.js";
+//
+//
+// export function winner(name) {
+//     const $winnerTitle = createNewElement('div', 'winnerTitle')
+//     if (name) {
+//         $winnerTitle.innerText = `${name} wins!`
+//         return $winnerTitle
+//     }
+//     $winnerTitle.innerText = 'draw'
+//     return $winnerTitle
+// }
