+++ helpers/helth_points/getIIPoints.js
@@ -1,4 +1,5 @@
+/*
 import {getRandomNumber} from "../getRandomNumber.js";
 import {ATTACK} from "../../store/body_parts_to_attack.js";
 
-export const getIIPoints = () => ATTACK[getRandomNumber(3) - 1]
+export const getIIPoints = () => ATTACK[getRandomNumber(3) - 1]*/
