+++ helpers/logs_and_banners/startBanner.js
@@ -1,5 +1,5 @@
-import {generateLogs} from "./generateLogs.js";
-
-export function startBattle(player1, player2) {
-    generateLogs('start', player1.name, player2.name)
-}
+// import {generateLogs} from "./generateLogs.js";
+//
+// export function startBattle(player1, player2) {
+//     generateLogs('start', player1.name, player2.name)
+// }
